refactor: コンポーネントのフォルダ構造を見直し
