+++ app/routes/__.poc.pokemons.$pokemonId._index/route.tsx
@@ -7,7 +7,7 @@ import {
   CardHeader,
   CardTitle,
 } from '~/components/shadcn/ui/card';
-import { LoadingDots } from '~/components/shared/loading-dots/loading-dots';
+import { LoadingDots } from '~/components/shared/loading-dots';
 import type { Pokemon } from '../__.poc._index/types/pokemon';
 import type { Route } from './+types/route';
